requirements.txt
asgiref==3.3.1
cairocffi==1.2.0
CairoSVG==2.4.2
cffi==1.14.0
cssselect2==0.3.0
defusedxml==0.6.0
Django==3.0.4
django-cors-headers==3.2.1
django-resized==0.3.11
django-rest-framework==0.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
html5lib==1.0.1
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.4
pandas==1.0.4
Pillow==7.1.2
pkg-resources==0.0.0
psycopg2==2.8.5
pycparser==2.20
Pyphen==0.9.5
python-dateutil==2.8.1
pytz==2019.3
six==1.15.0
sqlparse==0.3.1
tinycss2==1.0.2
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.2.0
xlrd==1.2.0
